requirements.txt
Flask==3.0.0
google-generativeai==0.5.0
pymupdf==1.23.26
Pillow==10.2.0